+++ CHANGELOG.md
@@ -2,8 +2,6 @@
 
 All notable changes to the "VirtualLabs" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
-
 ## [Unreleased]
 
-- Initial release
+- Started project with blank web extension with TypeScript and Webpack having hello world command